% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector (min-max)}
\usage{
normalize(x)
}
\description{
Normalize a numeric vector (min-max)
}
\examples{
IndepAssoc:::normalize(c(0, 5, 10))
}
\keyword{internal}
