% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{print.IndepPositivity}
\alias{print.IndepPositivity}
\title{Print an IndepPositivity summary}
\usage{
\method{print}{IndepPositivity}(x, ...)
}
\arguments{
\item{x}{An \code{IndepPositivity} object from \code{check_positivity()}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print an IndepPositivity summary
}
